tests/wpt/tests/svg/painting/parsing/stroke-width-valid.svg

<?xml version="1.0" encoding="UTF-8"?>
<svg xmlns="http://www.w3.org/2000/svg"
     xmlns:h="http://www.w3.org/1999/xhtml"
     width="800px" height="600px">
  <title>SVG Painting: parsing stroke-width with valid values</title>
  <metadata>
    <h:link rel="help" href="https://svgwg.org/svg2-draft/painting.html#StrokeWidth"/>
    <h:meta name="assert" content="stroke-width supports the full grammar '&lt;length-percentage&gt;' and unitless."/>
  </metadata>
  <g id="target"></g>
  <h:script src="/resources/testharness.js"/>
  <h:script src="/resources/testharnessreport.js"/>
  <h:script src="/css/support/parsing-testcommon.js"/>
  <script><![CDATA[

test_valid_value("stroke-width", "0");
test_valid_value("stroke-width", "10");
test_valid_value("stroke-width", "1px");
test_valid_value("stroke-width", "calc(2em + 3ex)");
test_valid_value("stroke-width", "4%");
test_valid_value("stroke-width", "5vmin");
test_valid_value("stroke-width", "calc(50% + 60px)");
test_valid_value("stroke-width", "calc(3)");
test_valid_value("stroke-width", "calc(2 + 1)", "calc(3)");
test_valid_value("stroke-width", "calc(2 + (7 - 5))", "calc(4)");

  ]]></script>
</svg>
